@@ -1,36 +1,55 @@
 {
+  "name": "react_todo-app-with-api",
   "homepage": ".",
-  "name": "react-typescript-starter-pack",
   "version": "0.1.0",
+  "keywords": [],
+  "author": "Mate Academy",
+  "license": "GPL-3.0",
   "dependencies": {
-    "@testing-library/jest-dom": "^5.12.0",
-    "@testing-library/react": "^11.2.6",
-    "@testing-library/user-event": "^12.8.3",
-    "@types/jest": "^26.0.23",
-    "@types/node": "^12.20.10",
-    "@types/react": "^17.0.4",
-    "@types/react-dom": "^17.0.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "4.0.3",
-    "typescript": "^4.2.4",
-    "typescript-eslint-parser": "^22.0.0",
-    "web-vitals": "^1.1.1"
+    "@fortawesome/fontawesome-free": "^6.1.2",
+    "@types/react-transition-group": "^4.4.5",
+    "bulma": "^0.9.4",
+    "classnames": "^2.5.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "5.0.1",
+    "react-transition-group": "^4.4.5",
+    "resolve-url-loader": "^5.0.0"
+  },
+  "devDependencies": {
+    "@cypress/react": "^5.12.4",
+    "@cypress/webpack-dev-server": "^1.8.4",
+    "@mate-academy/cypress-tools": "^1.0.5",
+    "@mate-academy/eslint-config-react-typescript": "^1.0.12",
+    "@mate-academy/scripts": "^1.2.12",
+    "@mate-academy/students-ts-config": "*",
+    "@mate-academy/stylelint-config": "*",
+    "@types/node": "^17.0.45",
+    "@types/react": "^18.0.15",
+    "@types/react-dom": "^18.0.6",
+    "cross-env": "^7.0.3",
+    "cypress": "^9.5.3",
+    "eslint": "^7.32.0",
+    "eslint-plugin-cypress": "^2.11.2",
+    "gh-pages": "^3.2.3",
+    "mochawesome": "^7.1.3",
+    "mochawesome-merge": "^4.2.0",
+    "mochawesome-report-generator": "^6.2.0",
+    "sass": "^1.70.0",
+    "stylelint": "^13.13.1",
+    "typescript": "^4.6.3"
   },
   "scripts": {
-    "lint": "eslint ./src --fix",
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
+    "start": "mate-scripts start -l",
+    "build": "mate-scripts build",
+    "test": "mate-scripts test -l",
+    "lint-js": "mate-scripts lint -j",
+    "lint-css": "mate-scripts lint -s",
+    "lint": "npm run lint-js && npm run lint-css",
+    "update": "mate-scripts update",
+    "postinstall": "npm run update && cypress verify",
     "predeploy": "npm run build",
-    "deploy": "gh-pages -d build"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
+    "deploy": "mate-scripts deploy"
   },
   "browserslist": {
     "production": [
@@ -44,10 +63,12 @@
       "last 1 safari version"
     ]
   },
-  "devDependencies": {
-    "@mate-academy/eslint-config-react-typescript": "^0.0.4",
-    "eslint": "^7.25.0",
-    "gh-pages": "^3.1.0",
-    "node-sass": "^5.0.0"
+  "mateAcademy": {
+    "_comment": "Replace 'reactTypescript' with 'react' if you want use React without Typescript",
+    "projectType": "reactTypescript",
+    "tests": {
+      "_comment": "Add `cypressComponents: true` to enable component tests",
+      "cypress": true
+    }
   }
 }
